Better Arbitrary Array instance
The Arbitrary Array instance now depends on the Gen's size parameter; if
the size is n, then the number of elements in a generated array should
be equally likely to be any number from 0 up to n-1.

This also fixes #35, a bug where vectorOf 0 x would contain exactly 2
elements rather than 0. This needed to be fixed to ensure that an
`arbitrary :: Gen (Array a)` is capable of producing empty arrays.